backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
shap==0.40.0
joblib==1.1.0
matplotlib==3.4.3
seaborn==0.11.2
requests==2.26.0
python-dotenv==0.19.1
cryptography==36.0.0